update the discovery
The Hue bridge periodically polls the portal, which stores both the external and internal (in the home network) IP of the bridge, plus the MAC address. If the portal has not received the “I’m alive” message from a specific bridge for a longer period it will consider the bridge disconnected. The Hue portal has linked the public IP address of the users network with the MAC address and local IP address of the Hue bridge. In case the 3rd party application runs from the same users network, and as such using the same public IP address, it is possible to retrieve the local IP address visiting the https://discovery.meethue.com/ webpage. This results in a reply with the list of connected bridges in the network. For example when three bridges are found the reply will be similar to: